simplify use of temporary directory
Only write the log file to local, not the final data file.  Writing the
data file remotely should be fine, and we would lose the local version
anyway if the network is down since condor deletes it.

This has the added advantage that the checkpoint file is simply a
modified version of the final output.